+++ LoopMeSDK/Internal/Network/LoopMeVideoManager.m
@@ -9,6 +9,7 @@
 #import <stdlib.h>
 #import "LoopMeVideoManager.h"
 #import "LoopMeError.h"
+#import "LoopMeGlobalSettings.h"
 
 NSInteger const videoLoadTimeOutInterval = 180;
 
@@ -17,9 +18,15 @@ NSInteger const videoLoadTimeOutInterval = 180;
     NSURLConnectionDataDelegate
 >
 
+@property (nonatomic, strong) id ETag;
+@property (nonatomic, strong) NSMutableURLRequest *request;
 @property (nonatomic, strong) NSURLConnection *connection;
 @property (nonatomic, strong) NSMutableData *videoData;
+@property (nonatomic, strong) NSTimer *videoLoadingTimeout;
+
 @property (nonatomic, strong) NSString *videoPath;
+@property (nonatomic, assign) long long contentLength;
+@property (nonatomic, assign, getter=isDidLoadSent) BOOL didLoadSent;
 
 - (NSString *)assetsDirectory;
 
@@ -49,18 +56,25 @@ NSInteger const videoLoadTimeOutInterval = 180;
     return [documentsDirectory stringByAppendingPathComponent:@"lm_assets/"];
 }
 
+- (void)invalidateTimers {
+    [self.videoLoadingTimeout invalidate];
+    self.videoLoadingTimeout = nil;
+}
+
 #pragma mark - Public
 
 - (void)loadVideoWithURL:(NSURL *)URL
 {
-    NSURLRequest *request = [NSURLRequest requestWithURL:URL cachePolicy:NSURLRequestReloadIgnoringLocalCacheData timeoutInterval:videoLoadTimeOutInterval];
-    self.connection = [NSURLConnection connectionWithRequest:request delegate:self];
+    self.videoLoadingTimeout = [NSTimer scheduledTimerWithTimeInterval:videoLoadTimeOutInterval target:self selector:@selector(timeOut) userInfo:nil repeats:NO];
+    self.request = [NSMutableURLRequest requestWithURL:URL cachePolicy:NSURLRequestReloadIgnoringCacheData timeoutInterval:videoLoadTimeOutInterval];
+    self.connection = [NSURLConnection connectionWithRequest:self.request delegate:self];
 }
 
 - (void)cancel
 {
     [self.connection cancel];
     self.connection = nil;
+    [self invalidateTimers];
 }
 
 - (void)clearOldCacheFiles {
@@ -95,7 +109,10 @@ NSInteger const videoLoadTimeOutInterval = 180;
     NSURL *URL = [NSURL fileURLWithPath:dataPath];
     
     if([data writeToFile:dataPath atomically:NO]) {
-        [self.delegate videoManager:self didLoadVideo:URL];
+        if (!self.isDidLoadSent) {
+            [self.delegate videoManager:self didLoadVideo:URL];
+            self.didLoadSent = YES;
+        }
     } else {
         [self.delegate videoManager:self didFailLoadWithError:[LoopMeError errorForStatusCode:LoopMeErrorCodeWrirtingToDisk]];
     }
@@ -117,19 +134,45 @@ NSInteger const videoLoadTimeOutInterval = 180;
     NSURL *URL = [NSURL fileURLWithPath:dataPath];
     return URL;
 }
+
+- (void)failedInitPlayer: (NSURL *)url
+{
+    self.didLoadSent = NO;
+    [self loadVideoWithURL:url];
+}
+
+- (void)reconect {
+    [self.request setValue:self.ETag forHTTPHeaderField:@"If-Range"];
+    
+    NSLog(@"%@", [NSString stringWithFormat:@"bytes=%lu-%lld", (unsigned long)self.videoData.length, self.contentLength]);
+    
+    [self.request setValue:[NSString stringWithFormat:@"bytes=%lu-%lld", (unsigned long)self.videoData.length, self.contentLength] forHTTPHeaderField:@"Range"];
+    self.connection = [NSURLConnection connectionWithRequest:self.request delegate:self];
+}
+
+- (void)timeOut {
+    [self cancel];
+    NSError *error = [LoopMeError errorForStatusCode:LoopMeErrorCodeVideoDownloadTimeout];
+    [self.delegate videoManager:self didFailLoadWithError:error];
+}
+
 #pragma mark - NSURLConnectionDelegate
 
 - (void)connection:(NSURLConnection *)connection didReceiveResponse:(NSURLResponse *)response
 {
     if ([response respondsToSelector:@selector(statusCode)]) {
         NSInteger statusCode = [(NSHTTPURLResponse *)response statusCode];
-        if (statusCode != 200) {
-            [connection cancel];
-            [self.delegate videoManager:self didFailLoadWithError:[LoopMeError errorForStatusCode:statusCode]];
+        if (statusCode == 200) {
+            self.contentLength = [response expectedContentLength];
+            self.ETag = [[(NSHTTPURLResponse *)response allHeaderFields] valueForKey:@"ETag"];
+            self.videoData = [NSMutableData data];
             return;
         }
+        if (statusCode != 206) {
+            [connection cancel];
+            [self.delegate videoManager:self didFailLoadWithError:[LoopMeError errorForStatusCode:statusCode]];
+        }
     }
-    self.videoData = [NSMutableData data];
 }
 
 - (void)connection:(NSURLConnection *)connection didReceiveData:(NSData *)data
@@ -139,7 +182,12 @@ NSInteger const videoLoadTimeOutInterval = 180;
 
 - (void)connection:(NSURLConnection *)connection didFailWithError:(NSError *)error
 {
+    if (error.code == -1005) {
+        [self reconect];
+        return;
+    }
     [self.delegate videoManager:self didFailLoadWithError:error];
+    [self invalidateTimers];
 }
 
 - (void)connectionDidFinishLoading:(NSURLConnection *)connection
@@ -147,6 +195,7 @@ NSInteger const videoLoadTimeOutInterval = 180;
     [self cacheVideoData:[NSData dataWithData:self.videoData]];
     self.videoData = nil;
     self.connection = nil;
+    [self invalidateTimers];
 }